Add support to Secure Spring App with JWT
This commit contains a brief information about the changes introduced to
configure your Spring Boot project to work with JWT authentication.

- `pom.xml`: Include dependencies for H2, spring security, spring data and jjwt
- `application.yml`: Include the configuration to connect to database
- `Role` and `User`: This classes that are mapped to a table in the
database
- `Role` and `User` Repositories: This classes are to perform the CRUD
operations for `Role` and `User`
- `UserService`: is an interface that extends from UserDetailsService
which is an interface that spring security use to get the users by
username
- `UserServiceImpl`: Contains the code that performs the operations that
interact with the user, this is the final component that implements the
UserService interface
- `JwtTokenUtil`: This class is used to generate the JWT Token but also to
perform operations such token validation and username extraction
- `JwtTokenFiler`: Is an implementation of OncePerRequestFilter, this
means that everytime a request is received it will be executed the
action `doFilterInternal` this will decide if is an request with or
without authentication
- `WebSecurityConfig`: This is a configuration class, here are all the
configurations to have our endpoints secured, there a bunch of comments
there to understand what is doing every section of the code.
- `GlobalConfig`: Is a class created to avoid having circular
dependencies where you can declare other beans if required
- `GlobalExceptionHandler`: This is a global `ControllerAdvice` it's
function is basically to catch the exceptions of our application and
turn it into an standard output in this case like the following sample:

```
{
    "error": "Forbidden",
    "message": "",
    "path": "/user",
    "status": 403,
    "timestamp": "2021-05-19T00:59:25.945+00:00"
}
```

- `SystemDataInitializer`: Is a class that generates the first users,
for this example there are 2 users created `admin@email.com` with
password `admin123` and `user@email.com` with password `user123` it's
recomended that on production environment this users should be deleted
or inserted manually on the database, this is only for testing porpuses.

- `AuthController`: expose the methods to login and register new users,
when the login is successful a value is returned in the `Authorization`
header, this is the same value that should be sent on each request like
the following sample

```
curl --location --request POST 'localhost:8080/api/login' \
--header 'Content-Type: application/json' \
--data-raw '{
    "username": "admin@email.com",
    "password": "admin123"
}'
```

```
curl --location --request GET 'localhost:8080/adminOrUser' \
--header 'Authorization: Bearer
eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiIxLGFkbWluQGVtYWlsLmNvbSIsImlzcyI6ImV4YW1wbGUuaW8iLCJpYXQiOjE2MjEzODYwMTIsImV4cCI6MTYyMTk5MDgxMn0.7S60_Bqt5aYxaYQT71wMbBILIbmqGBzPqQrnRRRsVefCCXiWJFeCSN6MtqeOHW-kd2J7j7Qx4A6W1Wn0wW9HOQ'
```

I have included some samples to test, in this case there are 4 endpoints
defined on the `DemoController`, those endpoints are:

`/`: This is accesible even if you are not logged
`/admin`: Only ADMIN_ROLE has access
`/user`:  Only USER_ROLE has access
`/adminOrUser`: Either ADMIN_ROLE or USER_ROLE has access

To configure the access to your endpoint you need to include the
annotation `@PreAuthorize` like in the following example

```
@PreAuthorize("hasAnyAuthority('USER_ROLE', 'ADMIN_ROLE')")
// Mapping and Method definition
```